+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Bump Compose BOM to 2024.02.00
+* Made `KEY_RESULT` constants in `Fragment`s `internal` to remove a footgun where the constant was easily confused with `KEY_REQUEST`
 
 ## 10.0.4